Backout 784bb7c5ea9c (Bug 799407) for bustage

+++ layout/base/nsBidiPresUtils.cpp
@@ -2102,8 +2102,7 @@ bool nsBidiPresUtils::WriteLogicalToVisual(const PRUnichar* aSrc,
     }
   }
 
-  NS_ASSERTION(static_cast<uint32_t>(dest - aDest) == aSrcLength,
-               "whole string not copied");
+  NS_ASSERTION(dest - aDest == aSrcLength, "whole string not copied");
   return true;
 }
